refactor(pack): refactor domain exchange hubic information promotioncode
