+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/domain/domain.html
