+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/hubic/hubic.html
