+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/informations/informations.html
