+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/promotionCode/promotionCode.modal.html
